Tested workaround for certificate validation issues.
If a processor's security is too tight, commands sent to IFTTT would fail, due to an inability to validate the certificate. These changes work around the issue. The long-term fix will be to add an optional flag on the Simpl+ module to allow selecting whether to circumvent vlidation or not.